mcp/denodo/requirements.txt
fastmcp>=2.9.0
httpx>=0.28.1
python-dotenv>=1.0.0